Fix models and model specs
