+++ spec/factories/questionnaires.rb
@@ -1,14 +1,19 @@
 FactoryBot.define do
   factory :questionnaire do
-    version_id { SecureRandom.uuid }
+    id { SecureRandom.uuid }
     questionnaire_type { "monthly_screening_reports" }
     dsl_version { 1 }
-    questionnaire_version do
-      association :questionnaire_version, strategy: :create,
-                  id: version_id,
-                  dsl_version: dsl_version,
-                  questionnaire_type: questionnaire_type
-    end
+    is_active { false }
+    layout {
+      {
+        type: "group",
+        view_type: "view_group",
+        display_properties: {
+          orientation: "vertical"
+        },
+        item: []
+      }
+    }
   end
 end
