+++ spec/models/questionnaire_spec.rb
@@ -1,20 +1,16 @@
 require "rails_helper"
 
 RSpec.describe Questionnaire, type: :model do
-  it { is_expected.to belong_to(:questionnaire_version) }
-
-  describe "delegated methods" do
-    it { should delegate_method(:localized_layout).to(:questionnaire_version) }
-    it { should delegate_method(:layout).to(:questionnaire_version) }
-    it { should delegate_method(:id).to(:questionnaire_version) }
-  end
-
   describe "Validations" do
     subject { create(:questionnaire) }
     it "validates uniqueness" do
       should validate_uniqueness_of(:dsl_version)
         .scoped_to(:questionnaire_type)
         .with_message("has already been taken for given questionnaire_type")
+    end
+
+    it "allows only one active form per type" do
+
     end
   end
 
@@ -25,10 +21,34 @@ RSpec.describe Questionnaire, type: :model do
       expect(described_class.for_sync).to include(discarded_questionnaire)
     end
 
-    it "includes nested sync resources" do
-      _discarded_questionnaire = create(:questionnaire, deleted_at: Time.now)
+    it "includes only the active forms" do
 
-      expect(described_class.for_sync.first.association(:questionnaire_version).loaded?).to eq true
+    end
+  end
+
+  describe "#localized_layout" do
+    it "localized text attributes in items recursively" do
+      allow(I18n).to receive(:t!).with("test_translations.test_string").and_return "Test"
+      layout = {"item" => [
+        {"text" => "test_translations.test_string",
+         "item" => [
+           {"text" => "test_translations.test_string"},
+           {"text" => "test_translations.test_string"}
+         ]},
+        {"text" => "test_translations.test_string"}
+      ]}
+
+      localized_layout = {"item" => [
+        {"text" => "Test",
+         "item" => [
+           {"text" => "Test"},
+           {"text" => "Test"}
+         ]},
+        {"text" => "Test"}
+      ]}
+
+      questionnaire = build(:questionnaire, layout: layout)
+      expect(questionnaire.localized_layout).to eq(localized_layout)
     end
   end
 end
